+++ src/main/kotlin/com/faforever/userservice/domain/LoginLog.kt
@@ -14,11 +14,11 @@ import java.time.LocalDateTime
 @Table("login_log")
 data class LoginLog(
     @Id
-    val id: Int,
+    val id: Long,
     @Column("login_id")
-    val userId: Int,
+    val userId: Long,
     val ip: String,
-    val attempts: Int,
+    val attempts: Long,
     val success: Boolean,
     @CreatedDate
     val createTime: LocalDateTime = LocalDateTime.now(),
@@ -34,9 +34,9 @@ data class FailedAttemptsSummary(
 )
 
 @Repository
-interface LoginLogRepository : ReactiveCrudRepository<LoginLog, Int> {
-    fun findByUserIdAndIp(userId: Int?, ip: String): Mono<LoginLog>
-    fun findByUserIdAndIpAndSuccess(userId: Int?, ip: String, success: Boolean): Mono<LoginLog>
+interface LoginLogRepository : ReactiveCrudRepository<LoginLog, Long> {
+    fun findByUserIdAndIp(userId: Long?, ip: String): Mono<LoginLog>
+    fun findByUserIdAndIpAndSuccess(userId: Long?, ip: String, success: Boolean): Mono<LoginLog>
 
     @Query(
         "SELECT sum(attempts) as total_attempts, count(login_id) as accounts_affected, min(create_time) as first_attempt_at, max(update_time) as last_attempt_at " +
